:truck: Moved UI implementation to Container.
