+++ tsconfig.json
@@ -13,6 +13,7 @@
     "noEmit": false,
     "paths": {
       "@components/*": ["src/client/components/*"],
+      "@containers/*": ["src/client/containers/*"],
       "@constants/*": ["src/common/constants/*"],
       "@pages/*": ["src/pages/*"],
       "@config/*": ["config/*"],
